pyproject.toml
[project]
name = "report-agent"
version = "0.1.0"
description = "Multi-Agent Report Generator"
requires-python = ">=3.11"
dependencies = [
    "langchain-core>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    "markdown>=3.7",
    "openai>=1.50.0",
    "pdfplumber>=0.11.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
    "sentence-transformers>=3.0.0",
    "tavily-python>=0.5.0",
    "weasyprint>=62.0",
    "faiss-cpu>=1.8.0",
    "pypdf>=6.9.1",
]

[tool.uv]
# 애플리케이션 프로젝트 — 패키지로 빌드하지 않음
package = false

[dependency-groups]
dev = []
